docs: render guide content at runtime
The original proposal aimed to transpile markdown on build-time to spare
client's resources and run-time code complexity.

Later it it showed that we already have a runtime library for rendering
markdown directly into the DOM. We decided to is that instead as it
requires no new dependencies and the runtime performance hit shouldn't
be that big.